test(tile-debug): triangles calculation file
